# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/gear/helmets": {
      "get": {
        "description": "Retorna cascos activos con sus variantes, tallas y precio mínimo por tienda. Soporta múltiples filtros.",
        "operationId": "HelmetModelsController_findAll",
        "parameters": [
          {
            "name": "brandSlug",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "full_face",
                "modular",
                "open_face",
                "half",
                "off_road"
              ]
            }
          },
          {
            "name": "shellMaterial",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "polycarbonate",
                "fiberglass",
                "carbon_fiber"
              ]
            }
          },
          {
            "name": "closureType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "double_d_ring",
                "micrometric",
                "quick_release",
                "magnetic",
                "pinch_buckle"
              ]
            }
          },
          {
            "name": "visorPinlock",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pinlock_30",
                "pinlock_30_max_vision",
                "pinlock_70",
                "pinlock_70_max_vision",
                "pinlock_120",
                "pinlock_120_max_vision",
                "pinlock_evo",
                "pinlock_protectint",
                "included_unknown",
                "not_included",
                "not_compatible"
              ]
            }
          },
          {
            "name": "certification",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "dot",
                  "snell",
                  "ece_2206",
                  "ece_2205",
                  "fim"
                ]
              }
            }
          },
          {
            "name": "colorFamily",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "black",
                "white",
                "red",
                "blue",
                "green",
                "yellow",
                "orange",
                "pink",
                "purple",
                "grey",
                "silver",
                "gold",
                "brown",
                "carbon",
                "multicolor",
                "bronze",
                "hi_vis"
              ]
            }
          },
          {
            "name": "sunVisor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "intercomReady",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "minSafetyRating",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 5,
              "type": "number"
            }
          },
          {
            "name": "maxWeightGrams",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "includeDeleted",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de cascos"
          }
        },
        "summary": "Listar cascos",
        "tags": [
          "Helmets"
        ]
      }
    },
    "/api/gear/helmets/{id}": {
      "get": {
        "description": "Retorna el detalle completo de un casco con todas sus variantes y tallas.",
        "operationId": "HelmetModelsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del casco"
          },
          "404": {
            "description": "Casco no encontrado"
          }
        },
        "summary": "Obtener casco por ID",
        "tags": [
          "Helmets"
        ]
      }
    },
    "/api/admin/gear/helmets": {
      "get": {
        "description": "Igual que el endpoint público pero incluye campos internos (SKU, deletedAt). Usa `includeDeleted=true` para ver eliminados.",
        "operationId": "HelmetModelsAdminController_findAll",
        "parameters": [
          {
            "name": "brandSlug",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "full_face",
                "modular",
                "open_face",
                "half",
                "off_road"
              ]
            }
          },
          {
            "name": "shellMaterial",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "polycarbonate",
                "fiberglass",
                "carbon_fiber"
              ]
            }
          },
          {
            "name": "closureType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "double_d_ring",
                "micrometric",
                "quick_release",
                "magnetic",
                "pinch_buckle"
              ]
            }
          },
          {
            "name": "visorPinlock",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pinlock_30",
                "pinlock_30_max_vision",
                "pinlock_70",
                "pinlock_70_max_vision",
                "pinlock_120",
                "pinlock_120_max_vision",
                "pinlock_evo",
                "pinlock_protectint",
                "included_unknown",
                "not_included",
                "not_compatible"
              ]
            }
          },
          {
            "name": "certification",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "dot",
                  "snell",
                  "ece_2206",
                  "ece_2205",
                  "fim"
                ]
              }
            }
          },
          {
            "name": "colorFamily",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "black",
                "white",
                "red",
                "blue",
                "green",
                "yellow",
                "orange",
                "pink",
                "purple",
                "grey",
                "silver",
                "gold",
                "brown",
                "carbon",
                "multicolor",
                "bronze",
                "hi_vis"
              ]
            }
          },
          {
            "name": "sunVisor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "intercomReady",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "minSafetyRating",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 5,
              "type": "number"
            }
          },
          {
            "name": "maxWeightGrams",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "includeDeleted",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de cascos (vista admin)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Listar cascos",
        "tags": [
          "Admin — Helmets"
        ]
      },
      "post": {
        "operationId": "HelmetModelsAdminController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHelmetModelDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Casco creado correctamente"
          },
          "400": {
            "description": "Datos inválidos"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Crear modelo de casco",
        "tags": [
          "Admin — Helmets"
        ]
      }
    },
    "/api/admin/gear/helmets/{id}": {
      "get": {
        "description": "Incluye todos los campos internos y variantes eliminadas.",
        "operationId": "HelmetModelsAdminController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle completo del casco (vista admin)"
          },
          "404": {
            "description": "Casco no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Obtener casco por ID",
        "tags": [
          "Admin — Helmets"
        ]
      },
      "put": {
        "description": "Todos los campos son opcionales (actualización parcial).",
        "operationId": "HelmetModelsAdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHelmetModelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Casco actualizado"
          },
          "404": {
            "description": "Casco no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Actualizar modelo de casco",
        "tags": [
          "Admin — Helmets"
        ]
      },
      "delete": {
        "description": "Marca el casco como eliminado sin borrarlo de la base de datos. Restaurable con /restore.",
        "operationId": "HelmetModelsAdminController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Casco eliminado (soft delete)"
          },
          "404": {
            "description": "Casco no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Eliminar casco (soft delete)",
        "tags": [
          "Admin — Helmets"
        ]
      }
    },
    "/api/admin/gear/helmets/{id}/restore": {
      "post": {
        "description": "Revierte el soft delete, el casco vuelve al estado activo.",
        "operationId": "HelmetModelsAdminController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Casco restaurado"
          },
          "404": {
            "description": "Casco no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Restaurar casco eliminado",
        "tags": [
          "Admin — Helmets"
        ]
      }
    },
...
  },
  "info": {
    "title": "Moto Gear Picker API",
    "description": "API para selección y gestión de equipamiento de moto",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateHelmetModelDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "brandId": {
            "type": "string"
          },
          "helmetType": {
            "type": "string",
            "enum": [
              "full_face",
              "modular",
              "open_face",
              "half",
              "off_road"
            ]
          },
          "safetyRating": {
            "type": "number",
            "minimum": 1
          },
          "shellMaterial": {
            "type": "string",
            "enum": [
              "polycarbonate",
              "fiberglass",
              "carbon_fiber"
            ]
          },
          "shellSizes": {
            "type": "number",
            "minimum": 1
          },
          "weightGrams": {
            "type": "number",
            "minimum": 0
          },
          "visorAntiScratch": {
            "type": "boolean"
          },
          "visorAntiFog": {
            "type": "boolean"
          },
          "visorPinlock": {
            "type": "string",
            "enum": [
              "pinlock_30",
              "pinlock_30_max_vision",
              "pinlock_70",
              "pinlock_70_max_vision",
              "pinlock_120",
              "pinlock_120_max_vision",
              "pinlock_evo",
              "pinlock_protectint",
              "included_unknown",
              "not_included",
              "not_compatible"
            ]
          },
          "sunVisor": {
            "type": "boolean"
          },
          "sunVisorType": {
            "type": "string"
          },
          "intercomReady": {
            "type": "boolean"
          },
          "intercomDesignedBrand": {
            "type": "string"
          },
          "intercomDesignedModel": {
            "type": "string"
          },
          "removableLining": {
            "type": "boolean"
          },
          "washableLining": {
            "type": "boolean"
          },
          "emergencyRelease": {
            "type": "boolean"
          },
          "closureType": {
            "type": "string",
            "enum": [
              "double_d_ring",
              "micrometric",
              "quick_release",
              "magnetic",
              "pinch_buckle"
            ]
          },
          "certification": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "dot",
                "snell",
                "ece_2206",
                "ece_2205",
                "fim"
              ]
            }
          }
        },
        "required": [
          "name",
          "slug",
          "brandId",
          "helmetType"
        ]
      },
      "UpdateHelmetModelDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "brandId": {
            "type": "string"
          },
          "helmetType": {
            "type": "string",
            "enum": [
              "full_face",
              "modular",
              "open_face",
              "half",
              "off_road"
            ]
          },
          "safetyRating": {
            "type": "number",
            "minimum": 1
          },
          "shellMaterial": {
            "type": "string",
            "enum": [
              "polycarbonate",
              "fiberglass",
              "carbon_fiber"
            ]
          },
          "shellSizes": {
            "type": "number",
            "minimum": 1
          },
          "weightGrams": {
            "type": "number",
            "minimum": 0
          },
          "visorAntiScratch": {
            "type": "boolean"
          },
          "visorAntiFog": {
            "type": "boolean"
          },
          "visorPinlock": {
            "type": "string",
            "enum": [
              "pinlock_30",
              "pinlock_30_max_vision",
              "pinlock_70",
              "pinlock_70_max_vision",
              "pinlock_120",
              "pinlock_120_max_vision",
              "pinlock_evo",
              "pinlock_protectint",
              "included_unknown",
              "not_included",
              "not_compatible"
            ]
          },
          "sunVisor": {
            "type": "boolean"
          },
          "sunVisorType": {
            "type": "string"
          },
          "intercomReady": {
            "type": "boolean"
          },
          "intercomDesignedBrand": {
            "type": "string"
          },
          "intercomDesignedModel": {
            "type": "string"
          },
          "removableLining": {
            "type": "boolean"
          },
          "washableLining": {
            "type": "boolean"
          },
          "emergencyRelease": {
            "type": "boolean"
          },
          "closureType": {
            "type": "string",
            "enum": [
              "double_d_ring",
              "micrometric",
              "quick_release",
              "magnetic",
              "pinch_buckle"
            ]
          },
          "certification": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "dot",
                "snell",
                "ece_2206",
                "ece_2205",
                "fim"
              ]
            }
          }
        }
      },
...
    }
  }
}